ajusta repository para retornar as querys corretas com paginação

+++ src/main/java/com/processo/seletivo/repository/FotoPessoaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FotoPessoaRepository extends JpaRepository<FotoPessoa, Long> {
-    List<FotoPessoa> findByPessoaPesIdOrderByFpDataDesc(Long pessoaId);
+public interface FotoPessoaRepository extends JpaRepository<FotoPessoa, Integer> {
+    List<FotoPessoa> findByPessoaPesIdOrderByFpDataDesc(Integer pessoaId);
 }
