+++ src/main/java/com/processo/seletivo/repository/ServidorTemporarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
     @Repository
-    public interface ServidorTemporarioRepository extends JpaRepository<ServidorTemporario, Long> {
+    public interface ServidorTemporarioRepository extends JpaRepository<ServidorTemporario, Integer> {
 }
